First/Last Commit
I started on project 4, but I didn't get to do much, but it is ready for review. Though I added in 2 links to 2 style sheets and also the script was changed for the javascript that was given to me. I added in a fieldset and a form within that fieldset, also some labels and inputs.